Change type of CImage::m_gzImageFIle to gzFile
Passing gZFile* m_gzImageFIle to gzread, gzwrite and gzclose fails due to
first parameter being wrong type on gcc at least since 4.4.

Remove the cast of gzopen return value to different type than what it is.
I did not find zlib requiring this kind of cast in any zlib since 0.71
(which is initial commit on github), so I expect we do not need it.

This same fix seems to be in Debian package partimage-0.6.8-2.2.

Signed-off-by: Marko Kohtala <marko.kohtala@gmail.com>